+++ sdk/tables/Azure.Data.Tables/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bugs Fixed
 
 - Properly handle `GenerateSasUri` when the client is constructed with a connection string ([#23404](https://github.com/Azure/azure-sdk-for-net/issues/23404))
+- Fixed an exception when constructing the `TableClient` with a connection string where the table name is the same as the account name.
 
 ### Other Changes